[PATCH] Make "guilt push" match "quilt push" when the patch doesn't apply
Sometimes the reason why the patch doesn't apply is because it has
already been applied in mainline when rebasing a patch series.  It is
therefore more convenient if "guilt push" leaves the working tree
untouched if the patch fails to apply unless explicitly asked to force
apply a failing patch using the -f option.

This behavior matches how "quilt push" and "hg qpush" works, so it
also avoids unpleasant surprises for developers used to other patch
series tools.

Signed-off-by: "Theodore Ts'o" <tytso@mit.edu>
Signed-off-by: Josef 'Jeff' Sipek <jsipek@cs.sunysb.edu>