[PATCH] Guilt: Do not allow patch name duplication
Quilt maintain a series of patches with unique file names. we might be
able to use sha1 under guilt instead, but this dis-allowing is much
simpler and works now.

so exit with error as soon as we find out that the given patch already
exist under the current branch dir.

Signed-off-by: Yasushi SHOJI <yashi@atmark-techno.com>
Signed-off-by: Josef 'Jeff' Sipek <jsipek@cs.sunysb.edu>