[PATCH] {push,pop}: Add -n <number of patches> option
Josef,

One of the things I've found lacking in guilt is the ability to
push or pop multiple patches at a time. The patch below does
this by adding the "-n <num>" option to these.

quilt does it with just a number instead of a patch name, but I'm
too lazy to implement that so '-n' it was ;)

[jsipek: documentation update & misc cleanup]
Signed-off-by: David Chinner <dgc@sgi.com>
Signed-off-by: Josef 'Jeff' Sipek <jsipek@cs.sunysb.edu>