guilt: anchor three dash divider to EOL.
The match on "^---" was too broad, and taking in things like
PGP sigs, and random ascii-art as a match for the divider between
the patch and patch header. This anchors the regex to the end of
the line.  It will still match on the 1st "^---$" encountered,
and not the last, so if you have multiple "^---$", then you loose.

Some patches don't have any sort of header, so you still need a match
on the "^--- " found in the oldfile/newfile stanza of the patch itself.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>
Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>