[PATCH] guilt: patchbomb: add some patch formatting options to the CLI
Allow the suppression of sign-offs so that only the sign-offs in the committed
patch are included in the formatted patch.

Allow the patch subject prefix to be modified from "[PATCH n/m]" to a user
configurable string.

Both of these options are passed to the invocation of git format-patch, and the
syntax of the --subject-prefix command is the same as for the git command.

Signed-off-by: Dave Chinner <dchinner@redhat.com>
Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>