[PATCH] guilt: fix "from: xxx" filtering in the patch description
Commit 2cc8d35 ([PATCH] guilt: improve patch header handling) caused
a regression with patch header handling.  When using a hand written
patch description in the format:

> Frobnozzle: this is a patch subject
>
> From: Fred McNurk <fred@mcnurt.foo>

the old code filtered the From: line out of the git commit
description, while the new code leaves it in there. The result is
that when using guilt patchbomb, the From line is emitted twice in
the email that is sent - once from the author of the patch, and
again in the body of the commit message.

Fix this by unconditionally filtering the From: field out of the
constructed commit message that is used for git.

This patch was sent out via my local version of guilt that has this
patch applied, so the sending of this patch is demonstration that
this fix has been tested and works...

Signed-off-by: Dave Chinner <dchinner@redhat.com>
Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>