regression: stop checking for existence of refs directories
git 2.13 changed the behavior of the ref directory cleanup code, so that
empty directories are removed more agressively.  The previous commit by
Eric fixes the failures on git 2.13 and newer, but breaks the tests for
older versions.  The simplest way to make the tests work for all
versions, is to stop checking for the existence of ref/patches
subdirectories.

In the future (when git's refs API actually becomes a thing), we might
have to rewrite the ref checking code as well since we won't be
guaranteed that each ref will show up as a file in .git/refs/patches.

Signed-off-by: Josef 'Jeff' Sipek <jeffpc@josefsipek.net>